+++ torchpipe/csrc/core/include/event.hpp
@@ -91,6 +91,17 @@ class SimpleEvents {
     std::unique_lock<std::mutex> lk(mut);
 
     bool done = data_cond.wait_for(lk, std::chrono::milliseconds(timeout_ms),
+                                   [this] { return (num_task == ref_count); });  //
+    if (eptr_) {
+      std::rethrow_exception(eptr_);
+    }
+    return done;
+  }
+
+  bool WaitDanger(uint32_t timeout_ms) {
+    std::unique_lock<std::mutex> lk(mut);
+
+    bool done = data_cond.wait_for(lk, std::chrono::milliseconds(timeout_ms),
                                    [this] { return (num_task == ref_count) || eptr_; });  //
     if (eptr_) {
       std::rethrow_exception(eptr_);